Add multithreaded memory footprint tests
Extend test cases to measure memory footprint for multithreaded
execution.
Throw exception when memkind_malloc() cannot allocate memory in
MemoryManager constuctor.
Rewrite functionality of proc_stats.h to avoid allocating local buffers.

+++ test/allocator_perf_tool/Allocation_info.cpp
@@ -23,10 +23,11 @@
 */
 
 #include "Allocation_info.hpp"
+#include <numaif.h>
 
-float convert_bytes_to_mb(size_t bytes)
+double convert_bytes_to_mb(uint64_t bytes)
 {
-	return ((float)bytes) / (1024.0*1024.0);
+    return bytes / (1024.0 * 1024.0);
 }
 
 int get_numa_node_id(void* ptr)
